vectorizing computation
